Printer shuts down gracefully
I initially did not use this approach because the 'stopping' callback
wouldn't always be run after the shutdown completed. Shutdown seems like
a good idea because it waits for the threads to finish before exiting
them. Still, we have to explicitly stop the server after shutdown if the
server is still running