Remove app preloading
The manager preloads the app but the manager is a long running process
meaning the app could become stale as the environment changes.
Preloading makes sense for production applications but as specjour is a
development tool saving 10 secs on environment load doesn't outweigh the
consequences of having a stale app loaded into memory.