Preloading presses on past StandardError
Scopes in Rails 3 get evaluated during the eager loading process which
may result in an exception being thrown when the database doesn't exist.
The after_fork block will create the database but if we don't catch the
exception, we never get there. This attempts to solve that problem but
may not work in all cases. It may be a better idea to catch the
exception, execute the hook, then abort.