Check for stale pid on listener
If the computer shuts down suddenly, or the listener is KILL'd the pid
will continue to exist without the process running. Now, we check if the
process is running before determining if the listener has started. We
don't have to remove the pid file because the listener will overwrite it.